# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PlayOn futsal management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /seasons:
    get:
      operationId: listSeasons
      tags: [seasons]
      summary: List seasons
      parameters:
        - name: active
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Seasons list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Season"
    post:
      operationId: createSeason
      tags: [seasons]
      summary: Create a season (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeasonInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
  /seasons/{id}:
    get:
      operationId: getSeason
      tags: [seasons]
      summary: Get a season
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Season
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
        "404":
          description: Not found
    patch:
      operationId: updateSeason
      tags: [seasons]
      summary: Update a season (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeasonUpdate"
      responses:
        "200":
          description: Updated season
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"

  # ── PROGRAMS (unified browse surface) ───────────────────────────────────────
...
  /referrals/generate:
    post:
      operationId: generateReferralCode
      tags: [referrals]
      summary: Generate (201) or retrieve (200) the caller's referral code. Blocked when program is disabled.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Existing pending referral code returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
        "201":
          description: New referral code generated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Referral"
        "400":
          description: Referral program is currently disabled
  /referrals/my:
    get:
      operationId: getMyReferrals
      tags: [referrals]
      summary: Get the caller's referral history and active code
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Referral hub data
          content:
            application/json:
              schema:
                type: object
                properties:
                  code:
                    type: ["string", "null"]
                  link:
                    type: ["string", "null"]
                  rewardCreditCents:
                    type: integer
                  isEnabled:
                    type: boolean
                  completedCount:
                    type: integer
                  pendingCount:
                    type: integer
                  referrals:
                    type: array
                    items:
                      $ref: "#/components/schemas/Referral"
  /referrals/claim:
    post:
      operationId: claimReferralCode
      tags: [referrals]
      summary: Claim a referral code (one per user, new signups only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
      responses:
        "200":
          description: Code claimed, reward issued to referrer
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  rewardCreditCents:
                    type: integer
        "400":
          description: Invalid request or program disabled
        "404":
          description: Code not found
        "409":
          description: Code already claimed or user already has a referral attribution
...
components:
  schemas:
...
    Season:
      type: object
      required: [id, name, startDate, endDate, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SeasonInput:
      type: object
      required: [name, startDate, endDate]
      properties:
        name:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
    SeasonUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        isActive:
          type: boolean
...
    Referral:
      type: object
      required: [id, code, referrerId, status, rewardCreditCents, createdAt]
      properties:
        id:
          type: integer
        code:
          type: string
        referrerId:
          type: integer
        referredUserId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, completed]
        rewardCreditCents:
          type: integer
        redeemedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...